requirements.txt
Django==2.2.23
djangorestframework==3.12.2
ipython
psycopg2
psycopg2-binary
python-decouple==3.1
django-widget-tweaks==1.4.8
autopep8
pandas
pylint
xlrd
